refactor: collect types in their own modules
